feat: add hourly weather graph
