+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import MainPage from './components/MainPage'
-import RootFrame from './components/RootFrame'
+import RootFrame from './components/rootFrame/RootFrame'
 import { LocationProvider } from './contexts/Location'
 
 const App: React.FC = () => (
