+++ src/types/location.ts
@@ -1,3 +1,8 @@
-import { City } from 'ts-open-weather-map'
+export type Location = { lat: number; lon: number; name: string; state?: string; country?: string }
 
-export type Location = City | { lat: number; lon: number; name: string }
+export function getFullLocationName(location: Location): string {
+  let output = location.name
+  if (location.state) output += `, ${location.state}`
+  if (location.country) output += `, ${location.country}`
+  return output
+}
